usecase-service/requirements.txt
python-dotenv
langchain
langchain-community
langchain-openai
langchain-chroma
langchain-huggingface
langchain-together
unstructured
chromadb
openai
tiktoken
flask
flask-cors